Updates ifib to return arbitrary precision integers

+++ fibonacci.jl
@@ -5,7 +5,7 @@ fib(n::Int) = n < 3 ? 1 : fib(n - 2) + fib(n - 1)
 # Iterative implementation
 
 function ifib(n)
-  j, k = 1, 1
+  j::BigInt, k::BigInt = 1, 1
   for l = 3:n
     j, k = k, j + k
   end
